pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "ics-ag-actor"
description = "ICS actor for the Auto Guider"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "PFS Collaboration"}
keywords = ["astronomy", "spectrograph", "subaru", "pfs", "autoguider"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Astronomy",
]
dependencies = [
    "astropy >= 5.0",
    "fitsio",
    "numpy >= 2.0",
    "psycopg2-binary",
    "sdss-opscore",
    "sdss-actorcore",
    "pfs-utils",
]
dynamic = ["version"]

[project.urls]
"Homepage" = "https://github.com/Subaru-PFS/ics_agActor"
"Bug Tracker" = "https://github.com/Subaru-PFS/ics_agActor/issues"
"Changelog" = "https://github.com/Subaru-PFS/ics_agActor/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "isort",
    "mypy",
    "pydocstyle",
]

[tool.setuptools.packages.find]
where = ["python/"]
include = ["agActor", "agActor.*"]

[tool.lsst_versions]
write_to = "python/agActor/version.py"

[tool.black]
line-length = 110
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 110
known_first_party = ["agActor"]

[tool.pytest.ini_options]
addopts = "--import-mode=importlib"

[tool.pydocstyle]
convention = "numpy"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
